+++ Source/Krypton Toolkit/Krypton.Toolkit.Suite.Extended.Outlook.Grid/Classes/Custom Columns/KryptonDataGridViewTokenColumn.cs
@@ -29,9 +29,11 @@ namespace Krypton.Toolkit.Suite.Extended.Outlook.Grid
         public KryptonDataGridViewTokenColumn()
             : base()
         {
-            this.CellTemplate = new TokenCell();
-            this.DefaultCellStyle.Alignment = DataGridViewContentAlignment.MiddleCenter;
-            this.ValueType = typeof(TokenCell);
+            CellTemplate = new TokenCell();
+            
+            DefaultCellStyle.Alignment = DataGridViewContentAlignment.MiddleCenter;
+            
+            ValueType = typeof(TokenCell);
         }
     }
 }