+++ Source/Krypton Toolkit/Krypton.Toolkit.Suite.Extended.Outlook.Grid/Classes/Custom Columns/KryptonDataGridViewTreeTextCell.cs
@@ -25,7 +25,7 @@ namespace Krypton.Toolkit.Suite.Extended.Outlook.Grid
     {
         private const int INDENT_WIDTH = 20;
         private const int INDENT_MARGIN = 5;
-        private Padding defaultPadding;
+        private Padding _defaultPadding;
         // private int glyphWidth;
 
         /// <summary>
@@ -34,7 +34,7 @@ namespace Krypton.Toolkit.Suite.Extended.Outlook.Grid
         public KryptonDataGridViewTreeTextCell()
             : base()
         {
-            defaultPadding = Style.Padding;
+            _defaultPadding = Style.Padding;
         }
 
         /// <summary>
@@ -53,15 +53,7 @@ namespace Krypton.Toolkit.Suite.Extended.Outlook.Grid
         /// <value>
         /// The glyph margin.
         /// </value>
-        protected virtual int GlyphMargin
-        {
-            get
-            {
-                return ((Level - 1) * INDENT_WIDTH) + INDENT_MARGIN;
-            }
-
-
-        }
+        protected virtual int GlyphMargin => ((Level - 1) * INDENT_WIDTH) + INDENT_MARGIN;
 
         /// <summary>
         /// Gets the level.
@@ -79,7 +71,9 @@ namespace Krypton.Toolkit.Suite.Extended.Outlook.Grid
                     return row.NodeLevel + 1; //during calculation 0 level must be 1 for multiplication
                 }
                 else
+                {
                     return -1;
+                }
             }
         }
 
@@ -96,8 +90,8 @@ namespace Krypton.Toolkit.Suite.Extended.Outlook.Grid
             int plus = 0;
             //if (hasChildNodes)
             //    plus = 15;
-            Style.Padding = new Padding(defaultPadding.Left + (level * INDENT_WIDTH) + INDENT_MARGIN + plus,
-                defaultPadding.Top, defaultPadding.Right, defaultPadding.Bottom);
+            Style.Padding = new Padding(_defaultPadding.Left + (level * INDENT_WIDTH) + INDENT_MARGIN + plus,
+                _defaultPadding.Top, _defaultPadding.Right, _defaultPadding.Bottom);
 
 
         }
@@ -329,9 +323,6 @@ namespace Krypton.Toolkit.Suite.Extended.Outlook.Grid
         /// <value>
         /// The owning node.
         /// </value>
-        public OutlookGridRow OwningNode
-        {
-            get { return base.OwningRow as OutlookGridRow; }
-        }
+        public OutlookGridRow OwningNode => base.OwningRow as OutlookGridRow;
     }
 }