+++ Documents/Help/Changelog.md
@@ -1,6 +1,8 @@
 =======
 
 ## 2022-04-04 - Build 2204 - April 2022
+* Resolved [#268](https://github.com/Krypton-Suite/Extended-Toolkit/issues/268), Crashes using Outlook Grid control
+* Resolved [#226](https://github.com/Krypton-Suite/Extended-Toolkit/issues/226), `KryptonOutlookGrid` throws a `MissingManifestResourceException` when trying to configure its columns
 * The `ThemeSelector` now uses the native `KryptonThemeComboBox` control
 * Nightly packages are now available both on GitHub and nuget.org
 * Reworked the `Toast` module (RTL support coming soon!)